SA-2209 - Move component to global section, adjust tests accordingly
