+++ src/components/reportBuilder/index.js
@@ -2,3 +2,4 @@ export { default as ActiveFilters } from './ActiveFilters';
 export { default as AggregatedMetrics } from './AggregatedMetrics';
 export { default as CompareByAggregatedMetrics } from './CompareByAggregatedMetrics';
 export { ChartGroups, Charts } from './Charts';
+export { IndustryBenchmarkModal } from './IndustryBenchmarkModal';
